More code cleanup - mainly on un-called functions.

+++ modem.h
@@ -45,7 +45,7 @@ extern void     delay5ms(uint16_t cnt);
 extern void     TL_module_first_run(void);
 extern void     check_led_type(void);
 extern uint8_t  read_ee(uint8_t page,uint8_t addr);
-extern void     write_ee(uint8_t page,uint8_t addr,uint8_t data_p);
+extern void     write_ee(uint8_t page, uint16_t addr, uint16_t data_p);
 extern void     set_sms_init(void);
 extern void     check_event(void);
 extern uint8_t  check_sms(void);
